pyproject.toml
[project]
name = "criminology-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.4",
    "django-ratelimit>=4.1.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.14.1",
    "python-dotenv>=1.1.1",
]
prod = [
    "celery>=5.5.3",
    "django-celery-beat>=2.8.1",
    "gunicorn>=23.0.0",
    "redis>=6.2.0",
]
